optimised code, implied both forms in one page and file, did few css changes
